Have softhsm create a temp. dir when RunSoftHSMSetup is called
Convert the softhsm support to a class and have it create a
temporary directory for storing the softhsm state in.

We need to set the environment variable SOFTHSM2_CONF when accessing
the module for it to find our custom config file location. So we set
this environment variable on the Pkcs11URI which will then be used
when the module is accessed.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>