pkcs11: Relax requirements on PIN for public key access
Public key access may not require a PIN, so do not report an error
if the PIN was not provided. Adjust a test case to deactivate the
PIN in the URI to test it.

For private key operations we require a PIN and if an error
occurred getting the PIN we report an error.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>