pkcs11: Use key id and/or label as search criteria
Neiher key id nor label need to be unique and the combination of them does
not need to be unique, either. So help pick a key use key id and/or label
as search criteria for the key. If the key id + label are not unique criteria
for identifying a key we still refuse to use the key.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>