pkcs11: Require 'token' attribute if 'slot-id' was not given
When the user did not provide a slot-id we walk the list of slot Ids
and use a session with the first one that accepts our PIN. This can
lead to using the wrong slot. Instead, require that the user provide
the 'token' attribute and use it to compare against the token label
of each slot and only log into a slot if the token label matches. If
there are multiple slots with the same token label we will use the first
one that accepts the PIN.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>