decryption: Accumulate error messages in case no key can decrypt
When decrypting we may try a lot of keys to find one that does the
decryption. Sometimes no key works and in this case it is important
to know what went wrong. So, let's accumulate the error messages in
those loops where we would otherwise just 'continue' in case an error
occurred.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>