pkcs11: Get PIN also in case of public key op but don't require it
Some devices also require a PIN when trying to find a public key, others
do not require it. So get the PIN if it is available but do not require
it to be available for public key related operations or finding a public
key.

The IBM grep11 module seems to now require a login with a PIN when trying
to find a public key object.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>