pkcs11: Query OCICRYPTO_OAEP_HASHALG env. variable when using pkcs11 pub key
Also query the environment variable OCICRYPTO_OAEP_HASHALG when using a
pkcs11 public key to give the user control over the hash algorithm to use.
Both grep11 and softhsm need 'sha1' and cannot use sha256 but softhsm
fails in EncryptInit() while the grep11 module fails later in Encrypt().

Since we now have 2 modules that require 'sha1', let's make sha1 the default
and sha256 an option. This label does not seem to be so security ciritical
that the hash would make a big difference.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>