pkcs11: Implement functions for loading and parsing a config file
Implement functions for loading and parsing an ocicrypt pkcs11 config file.
The config file may look like this:

pkcs11:
  module-directories:
  - /usr/lib64/pkcs11/
  allowed-module-paths:
  - /usr/lib64/pkcs11/libsofthsm2.so

The 'module-directories' enumerates directories in which to look for pkcs11
modules. The 'allowed-module-paths' restricts the pkcs11 modules that can
be used. The list may contain filenames (shown above) or directories, which
have to end with '/'. The 'allowed-module-paths' part must not be empty
since this is equivalent to 'deny-all'.

The config file is searched for in this order:

- If ${OCICRYPT_CONFIG}="internal" then use the internal allow-all policy.
- ${OCICRYPT_CONF}
- ${XDG_CONFIG_HOME}/ocicrypt-pkcs11.conf
- ${HOME}/.config/ocicrypt-pkcs11.conf
- /etc/ocicrypt-pkcs11.conf

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>