pkcs11: Get the user's pkcs11 config opts only if pkcs11 keys are used
Rather than passing the Pkcs11Config from the top down, get them only when
it is clear that a pkcs11 device is involved. This helps avoid unnecessary
error messages on the top level.

The Pkcs11Config is only needed when Pkcs11 Yaml files, that hold the pkcs11
URI, are given, so we only require it to be passed in this case. In case
of PEM files the Pkcs11Config is not needed since we will not use a pkcs11
module.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>