helpers: Modify parsing of private keys for key provider protocol
This patch modifies the parsing of the private keys in such a way that
we can keep the implementation of processPrivateKeyFiles implemented
in a separate file in imgcrypt in sync with the one in ocicrypt. In
particular, imgcrypt gets all private keys from the function
processPrivateKeyFiles that parses whatever was passed via the
command line with the '--key' option, so we now also parse the
identifier for the private key of a custom keyprovider in this
function as well, even though it may not be a 'file'.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>