travis: also test on Go 1.16.x, and "go mod tidy" with go 1.16
This adds the current (1.16) version to travis for testing. I kept the
existing version (go 1.13), to verify older versions of Go still function
correctly (but this could be removed at some point).

Also ran `go mod tidy` with go 1.16, as there have been changes in how
`go.sum` is generated, resulting in a diff.

Note that while CI does run "go mod tidy", it does not currently validate
`go.mod` and `go.sum` for changes, so would not catch such changes, which
may be a good addition to make in a follow-up.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>