pkcs11: Move to sha256 for the default hash to use
In case the user doesn't set the environment variable
OCICRYPT_OAEP_HASHALG sha256 will be used now. This breaks default
usage with SoftHSM because the only hash it currently (v2.6.1) supports
is sha1. So a user of SoftHSM now has to set the environment variable
to 'sha1' and we have to adjust the test case because of this.

SoftHSM link to OAEP only supporting sha1:
https://github.com/opendnssec/SoftHSMv2/blob/7f99bedae002f0dd04ceeb8d86d59fc4a68a69a0/src/lib/SoftHSM.cpp#L3123-L3127

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>