linters: Disable report of deprecated packages by staticcheck
Suppress the report of this type of warning sine this package is fine for
us:

SA1019: "golang.org/x/crypto/openpgp" is deprecated: this package is
unmaintained except for security fixes. New applications should consider
a more focused, modern alternative to OpenPGP for their specific task.
If you are required to interoperate with OpenPGP systems and need a
maintained package, consider a community fork.
See https://golang.org/issue/44226. (staticcheck)

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>